Bifid Cipher (now Deciphers)
Finally, I managed to reuse most of the code to cipher in the decrypting process.
It is not optimized and for some reason if I wanted to use the "getNewMessage(finalCoords, Matrix)" function for the decypher part, it wouldn't allow me, cause conflicting types with ints and strings. Will check on it later / another day / never.